+++ src/Exceptions/InvalidInputException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Wink\ModelGenerator\Exceptions;
+
+class InvalidInputException extends \RuntimeException
+{
+}